Fixed the awful awful awful bug =(
If you start to charge a core, but then leave before you finish charging it,
and then enemies reduce it down to 0 health, it counted as losing a core even
though you hadn't acquired it. This is a very common occurrence, and
essentially ruins the game.
* Music tracks are not played when they should be playing
* Tower inventory is permanently shrunk, handicapping the player
* The chain never opens, rendering the game unbeatable

so basically a softlock =(